# Conflicts:
#	.DS_Store
#	README.md
#	conf/httpd.conf
#	src/mod_ntlm.c
#	src/mod_ntlm_authentication.c
#	src/mod_ntlm_authorization.c
#	src/mod_ntlm_interface.c
#	src/mod_ntlm_version.h

+++ src/mod_ntlm_interface.c
@@ -247,7 +247,7 @@ static void sspi_set_domain(sspi_auth_ctx *ctx)
 		return;
 
 	ctx->hdr.Domain = ctx->crec->sspi_domain;
-	ctx->hdr.DomainLength = (unsigned long)strlen(ctx->crec->sspi_domain);
+	ctx->hdr.DomainLength = strlen(ctx->crec->sspi_domain);
 }
 
 static void sspi_set_default_domain(sspi_auth_ctx *ctx)
@@ -261,7 +261,7 @@ static void sspi_set_default_domain(sspi_auth_ctx *ctx)
 		return;
 
 	ctx->hdr.Domain = ctx->crec->sspi_default_domain;
-	ctx->hdr.DomainLength = (unsigned long)strlen(ctx->crec->sspi_default_domain);
+	ctx->hdr.DomainLength = strlen(ctx->crec->sspi_default_domain);
 }
 
 /* working up with SSPI header */
